Moved login input + logic up to container component
